Create read.me
COHORT ANALYSIS

What is Cohort Analysis?
A cohort represents a group of a population or an area of study which shares something in common within a specified period. For example, a group of people born in India in 2000 is an example of a cohort related to the number of births in a country. Likewise, in terms of business problems, cohorts represent a group of customers or users. For example:

Several users who purchased the subscription the app in a given period.
The number of users who cancelled a subscription during the same month.